pyproject.toml
[project]
name = "sanic-boot"
version = "0.0.4.1"
description = "A web framework base on Sanic"
readme = "README.md"
authors = [{ name = "嬴寒", email = "yinghan22@163.com" }]
requires-python = ">=3.10"
dependencies = [
    "sanic<=25.3.0",
    "tortoise-orm<=0.25.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true


[project.scripts]
sanic-boot-init = "sanic_boot.manager:sanicBootInit"
sanic-boot = "sanic_boot.manager:sanicBooter"
